Rename queries to operations
